+++ js/main.js
@@ -1,4 +1,4 @@
-import {changeScreen} from './until';
+import {changeScreen} from './util';
 import introScreen from './intro-screen';
 
 changeScreen(introScreen);
